Pandemic Plan3.0

* Fix-Fixing errors in offers  photos-PIIB12021-277View offers-Diego Barquero,Isaac Herrera

* Fix redirect link

* Fix paymethod page name

* Merge branch 'Pandemic_Plan3.0' into OfferPhotos

* Add link to payment methods

* Merge branch 'Pandemic_Plan3.0' into PandemicEntrepreneurPayMethod

* Feat-Adding offer service photos-PIIB12021-277View offers-Diego Barquero,Isaac Herrera

* Merge branch 'Pandemic_Plan3.0' of https://bitbucket.org/cristian_quesadalopez/ecci_ci0128_i2021_g01_pi into Pandemic_Plan3.0

* solving conflicts

* Solving merge conflicts

Approved-by: hellen.fuentesartavia
Approved-by: jorge.mayorgaulate
Approved-by: Silvia Aguilar
Approved-by: luis.rojascarrillo

+++ EmprendeUCR/EmprendeUCR_WebApplication/Data/Services/Categories/DeleteCategoryService.cs
@@ -6,6 +6,7 @@ using Microsoft.AspNetCore.Mvc.RazorPages;
 using Microsoft.EntityFrameworkCore;
 using EmprendeUCR_WebApplication.Data.Entities;
 using Syncfusion.Blazor.TreeGrid;
+using EmprendeUCR_WebApplication.Data.Services;
 
 namespace EmprendeUCR_WebApplication.Data.Services.Categories
 {
@@ -15,15 +16,16 @@ namespace EmprendeUCR_WebApplication.Data.Services.Categories
         SfTreeGrid<Category> TreeGrid;
         public DeleteCategoryService(Contexts.SqlServerDbContext context)
         {
-            _context = context;
             ResetRemoveCategoryData();
+            _context = context;
         }
-
         /*
          * 
          * Private Variables
          * 
          */
+        Product_ServiceService p_s_service;
+        private bool deleted = false;
         private int SelectedId = -1;
         private Category SelectedCategory;
 
@@ -32,6 +34,7 @@ namespace EmprendeUCR_WebApplication.Data.Services.Categories
          * Public Variables
          * 
          */
+        public bool deleteNotValid = false;
         public string Title = null;
         public bool RemoveDialogVisible;
         public bool RemoveToolbarItemDisabled;
@@ -56,6 +59,8 @@ namespace EmprendeUCR_WebApplication.Data.Services.Categories
 
         private void ResetRemoveCategoryData()
         {
+            deleteNotValid = false;
+            deleted = false;
             RemoveToolbarItemDisabled = true;
             Title = null;
             SelectedId = -1;
@@ -65,26 +70,61 @@ namespace EmprendeUCR_WebApplication.Data.Services.Categories
         {
             this.TreeGrid = main;
             this.RemoveDialogVisible = false;
-            var SelectedIndex = await TreeGrid.GetRowIndexByPrimaryKey(SelectedCategory.Id);
-            await TreeGrid.DeleteRecord("Nombre",SelectedCategory);
             RemoveCategory(SelectedId);
+            if (!deleted)
+            {
+                var SelectedIndex = await TreeGrid.GetRowIndexByPrimaryKey(SelectedCategory.Id);
+                await TreeGrid.DeleteRecord("Nombre", SelectedCategory);
+            }
         }
         public void RemoveCategory(int Id)
         {
-           RemoveCategoryRecursive(Id);
+            Console.WriteLine("Llega al recursivo");
+            RemoveCategoryRecursive(Id);
+            
         }
         public void RemoveCategoryRecursive(int Id)
         {
             IList<Category> listChilds =_context.Category.Where(c => c.ParentId.Equals(Id)).ToList();
             foreach(Category category in listChilds)
             {
-                 RemoveCategoryRecursive(category.Id);
+                if (AnyProductsServicesAssociate(category.Id) && (!deleted))
+                {
+                    RemoveCategoryRecursive(category.Id);
+                }
+                else
+                {
+                    deleted = true;
+                }
+            }
+            if (!AnyProductsServicesAssociate(Id))
+            {
+                deleted = true;
+            }
+            if (!deleted)
+            {
+                Category CategoryToRemove = _context.Category.Find(Id);
+                _context.Category.Remove(CategoryToRemove);
+                _context.SaveChanges();
+                TreeGrid.DeleteRecord(Title, SelectedCategory);
+            }
+            else
+            {
+                deleteNotValid = true;
             }
             
-            Category CategoryToRemove =  _context.Category.Find(Id);
-            _context.Category.Remove(CategoryToRemove);
-            _context.SaveChanges();
-            TreeGrid.DeleteRecord(Title, SelectedCategory);
+        }
+        public bool AnyProductsServicesAssociate(int categoryId)
+        {
+            bool any = true;
+            var products = _context.Product_Service.Where(p => p.Category_ID.Equals(categoryId)).ToList().Length();
+            Console.WriteLine("Esta es la cantidad de asociados: ");
+            Console.WriteLine(products);
+            if (products != 0)
+            {
+                any = false;
+            }
+            return any;
         }
     }
 }
